requirements.txt
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
filetype==1.2.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.0
minify_html==0.18.1
multimethod==1.12
networkx==3.6.1
numba==0.62.1
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
phik==0.12.5
pillow==12.0.0
psycopg2-binary==2.9.11
puremagic==1.30
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
seaborn==0.13.2
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
statsmodels==0.14.6
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
visions==0.8.1
webdriver-manager==4.0.2
websocket-client==1.9.0
wordcloud==1.9.5
wsproto==1.3.2
ydata-profiling==4.18.0
